# Conflicts:
#	postchain-base/src/main/kotlin/net/postchain/service/PostchainService.kt
#	postchain-cli/src/test/kotlin/net/postchain/cli/CliIntegrationIT.kt

+++ postchain-gtv/src/test/kotlin/net/postchain/gtv/merkle/MerkleHashCalculatorDummy.kt
@@ -15,7 +15,7 @@ import java.nio.charset.Charset
 fun dummySerializatorFun(iGtv: Gtv): ByteArray {
     when (iGtv) {
         is GtvInteger -> {
-            val i: Long = iGtv.integer.toLong()
+            val i: Long = iGtv.integer
             if (i > 127 && i > -1) {
                 throw IllegalArgumentException("Test integers should be positive and should not be bigger than 127: $i")
             } else {
